feat(#323): level and reverse shape the isolated looper playback
export_mixdown now applies mix (level) and reverse (in addition to decay); a
content revision on the slot bumps on every change that alters the mixdown
(record close, overdub, undo/redo, clear, level, decay, reverse). The
controller re-arms the isolated stream only when that revision moves, so the
level/reverse controls take effect without respawning a render on a steady
loop. Speed remains follow-up (a rate change belongs to the stream, not a
buffer rewrite). Workspace 3909 green.

+++ crates/infra-cpal/src/controller.rs
@@ -130,7 +130,7 @@ pub struct ProjectRuntimeController {
     /// #323: the loop content `(len_frames, layers)` each looper's isolated
     /// stream was last armed with, so `sync_looper_streams` re-arms only when
     /// the recording actually changed — not every meter tick.
-    pub(crate) looper_armed: RefCell<HashMap<(ChainId, u64), (usize, usize)>>,
+    pub(crate) looper_armed: RefCell<HashMap<(ChainId, u64), (u64, u64)>>,
     /// Single owner of every jackd process openrig controls on Linux. Replaces
     /// the former ensure_jack_running / stop_jackd_for / jack_meta_for set of
     /// free functions with an explicit state machine (issue #308).
